Remove ServerConfig from QUICRelayProtocol class
